data generation seprated & ntrace_tot in set_manual fixed (ctfile.py)
